Find the root cell

Use either the primary manifold embedding of cell similarity to find the root cell by cluster-cell rank-correlation or use the flattened representation of this embedding to find the root cell by maximum separation heuristic. Possible values for cluster_order_by are predecessor and distance. Possible values for cell_order_by are index and distance.

FindRootCell(
  use_flat_dist = TRUE,
  cluster_order_by = "distance",
  cell_order_by = "distance",
  graph_cluster_mst = NULL,
  dist_graph = NULL,
  dist_flat,
  cluster_labels = NULL,
  root_cluster = NULL
)

Arguments

use_flat_dist
use max path length heuristic on flat embedding

cluster_order_by
cluster rank parameter for measuring cluster/cell rank correlation

cell_order_by
cell-wise pseudotime parameter for measuring cluster/cell rank correlation

graph_cluster_mst
igraph object representing an mst on the cluster-cluster graph

dist_graph
a distance matrix of cells embedded in a graph

dist_flat
the manifold embedding of the cells

cluster_labels
the cluster label for each cell

root_cluster
the id of the root cluster on the cluster-cluster graph

Value

integer index of the root cell