Get the marker genes for each cluster

Get the marker genes for each cluster and report their 1-normalized expression value.

GetMarkerTable(counts_data, cluster_labels, H = NULL, n_sorted = 50,
  gene_names, gene_expression_threshold = NULL, n_features = NULL,
  use_H = FALSE)

Arguments

counts_data
a matrix of expression values for each cell (rows) and gene (columns) which will be normalized and

cluster_labels
a vector of cluster labels

H
a nonnegative matrix such that W = H*t(H), H_i,j is the cells_weight by which cell i belongs to the jth cluster

n_sorted
the number of markers to report in the n_sorted table

gene_names
gene symbols corresponding to rows in counts_data (this should match rows of M, if provided)

gene_expression_threshold
for n cells, for gene_expression_threshold = m, dont consider genes expressed in more than n-m cells or genes expressed in less than m cells

n_features
number of marker genes per cluster to retrieve

use_H
whether to use H as loading, default is false, instead using cumulative absolute difference to sort max-mean assigned labels

Value

a list containing

all

all the markers listed by gene index, cluster, and score

n_sorted

a subset of the sorted marker table with gene symbols