Read Counts Data

Read counts data from file and annotate the generated Single Cell Experiment Object. Right now this function assumes: - mouse cells - only spike ins are ERCC - row and column names are provided in the data - the row names are gene symbols In the case of GEO data, it takes a GSE accession code and assumes that the supp files contain all the expression data

ReadAnnotateCountsSCE(filepath = NULL, inf_exp_rate = 0,
  inf_gene_count = 0, inf_cell_count = 0, series_accession,
  metad_title = "characteristics_ch1", cell_id = "title",
  pheno_filter = NULL, plot_title = NULL, GEOfilecache = getwd(),
  annotation_db, feature_annotation_key, remove_spike_ERCC = FALSE,
  pre_clust = FALSE, rescale = TRUE, var_gene_method = "mean_var",
  ...)

Arguments

filepath
the full or relative (to the workspace) path of the data file

inf_exp_rate
the minimum mean expression rate of retained genes

inf_gene_count
retain a gene if more than this number of cells express it

inf_cell_count
retain a cell if more than this number of genes are expressed in it

series_accession
the series accession code, eg "GSE67602"

metad_title
the series accession code, eg "GSE67602"

cell_id
the series accession code, eg "GSE67602"

pheno_filter
a vector of strings to partially match cell metadata

plot_title
path of the saved mean/variance plot, default NULL, if NULL the plot will be printed instead

GEOfilecache
location for caching series supplemental files, default is current directory

annotation_db
annotation package for the data, eg "org.Mm.eg.db"

feature_annotation_key
the key for annotation.db lookups, eg "SYMBOL"

remove_spike_ERCC
remove ERCC spike ins after they are used for normalization

pre_clust
whether or not to precluster the cells before getting size factors

rescale
whether to compute size factors and rescale counts

var_gene_method
how to select variable genes, eg 'mean_var', 'pca'. mean_var selects the genes whose variance is highest after removing the fitted spike-in-based technical variance. pca finds the genes with the max rotation coefficients

...
arguments to called functions

Value

a list containing:

c(seur, sce)

a seurat or sce object