Generate multiple representations

Generate representations of the data to be used in lineage analysis.

RepresentationMap(flat_embedding = NULL, similarity_matrix,
  join_components = TRUE, normalize_S = TRUE,
  flat_embedding_method = "umap", ...)

Arguments

flat_embedding
optionally provided low dim embedding, if not then 2d tsne will be used

similarity_matrix
the graphical embedding of the cells

join_components
boolean, whether or not to join disconnected components of the similarity matrix

normalize_S
whether or not to normalize the similarity matrix

flat_embedding_method
what method to use for flat embedding. Use umap by default. Can be umap or tsne.

...
arguments to called functions

Value

a list containing:

dist_flat

distance matrix on flat embedding

dist_graph

distance matrix on graph

adj_matrix

unweighted adjacency matrix

flat_embedding

the low dimensional embedding of the cells

similarity_graph

igraph object on the unweighted adjacency matrix

components

an igraph components object based on similarity_graph

n_components

number of components in components

sizes

sizes of components in components

members

list of members of components, sorted by component size and member index